refactor: move deck import view to feature
