+++ src/pages/deck-import/ui/DeckImportPage.tsx
@@ -5,10 +5,8 @@ import { useKey } from "react-use";
 import { createDeck, useDecks } from "@/entities/deck";
 import { createCard, editCard, generateCardId, useCards } from "@/entities/card";
 import { usePreferences } from "@/entities/preferences";
-import { downloadSampleCsv, SAMPLE_CSV_TEXT, useDeckImport } from "@/features/deck-import";
+import { DeckImportView, downloadSampleCsv, SAMPLE_CSV_TEXT, useDeckImport } from "@/features/deck-import";
 import { AppLayout } from "@/widgets/app-layout";
-
-import { DeckImportView } from "./DeckImportView";
 
 export const DeckImportPage: React.FC = () => {
   const preferences = usePreferences();
